Renderer: Add a renderer class
It will handle rendering drawables instead of Scene.
Inside the Renderer we create an unordered map of shaders
and all drawables that use specific ones. This enables us
to group objects that use the same shaders together, and
thus limiting the amount of GPU memory shuffling.

We can probably be even more sophisticated in the future.