Renderer: Don't keep drawables as weak_ptrs+don't keep instanced drawables
weak_ptrs require a lock() call which is not particulary cheap,
so it's better not to call it in something performance intensive
like a render loop... also we probably won't have proper memory
disallocation for a long time.

Instanced drawables are already kept in Material so there is no need
to duplicate them in Renderer.