Everywhere: Use Badge pattern for (every?) use of factory pattern
Because of the way std::make_shared works, we couldn't mark constructors
as private in the case where we had factory functions producing
shared_ptrs to objects, that were created using the std::make_shared
function. This function has a nice optimization described ex. here:
https://medium.com/@its.me.siddh/modern-c-series-shared-ptr-vs-make-shared-when-to-use-what-7b2281893927

Keeping constructors public meant that the programmer might use it
instead of the desirable create() functions.
By using the Badge pattern we get ~both pieces of the cake.