ResourceManager: Add implementation for Texture, Mesh and Shader
ResourceManager handles sharing of resources that can be shared
(like the ones mentioned above). This brings us memory improvements
(because we don't need to store multiples of the same texture)
and performance improvements because we don't need to compile shaders
each time one is created, if they are the same as ones already loaded
(and we currently don't load a lot of different shaders).

ResourceManager currently doesn't allow any loaded resources to be
unloaded by keeping a shared_ptr to them. This allows us to instantly
load shaders and models, even if they are not currently used by any
entity for a brief moment. This also means that if someone would
create hundreds of spheres (generated by us) with different sector
or stack counts, none of them would unload. But this scenario is
currently not present.