Entity: Add methods to create internal entities and components
Internal entities and components do not belong to any scene,
they won't be awaken, started or updated. They might be rendererd.
Internal entity + Camera component will be used for the editor
camera. Having them behave as an entity allows us to easily
reuse already exisiting Transform and Camera component code.

This is not a good solution but the best I could find.