Component: Add reprepare() method
This method will be used by the deserialization code, to "reprepare"
a component after creating it (which usually "prepares" the component)
and assigning appropriate values to all serialized member variables.

This method was already used in Model and its children.

This is not a perfect solution, but it was the only reasonable that
came to my mind.