RM+Scene: Clear all materials from shaders after unloading a scene
When (re)loading a scene, meshes, textures and shaders
are preserved so we can load everything faster. There is a
caveat that shaders hold onto shared_ptrs of their materials,
and materials are currently not managed by RM, but are recontstructed
each time a scene is loaded. This means that we accumulate materials
from previous loads. As a hack, let's just clear those materials
when unloading a scene.