Drawable+Light: Don't reregister components in on_enabled
These components are already registered at the beginning
in create() or initialize() methods. When first starting
the game, we don't want to reregister them. This leads
to bugs, crashes and performance regressions.

Note that due to a bug we don't actually call on_enabled
at the beginning of the game, which is why these things
were not seen. This will change in the next commit.