Entity+SceneSerializer: Call awake() after deserialization ends
Previously when deserializing entities while the game was running
(loading prefabs) we would call awake() and on_enable() immediately
after adding a component to an entity. Other components would not
have been deserialized yet, at this point, so anything in these
methods that would reference values of variables in those other
components, would read zero initialized data, and not the data
that was serialized.